Moves Core Data models to their own files.
