+++ FeedStoreChallenge/CoreData/CoreDataFeedStore.swift
@@ -67,56 +67,3 @@ public final class CoreDataFeedStore: FeedStore {
         context.perform { action(context) }
     }
 }
-
-// MARK: - Core Data Models
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-    @NSManaged public var timestamp: Date
-    @NSManaged public var feed: NSOrderedSet
-    
-    var localFeed: [LocalFeedImage] {
-        return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
-    }
-    
-    @nonobjc public class func fetchRequest() -> NSFetchRequest<ManagedCache> {
-        return NSFetchRequest<ManagedCache>(entityName: "ManagedCache")
-    }
-    
-    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request: NSFetchRequest<ManagedCache> = fetchRequest()
-         request.returnsObjectsAsFaults = false
-         return try context.fetch(request).first
-     }
-    
-    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-         try find(in: context).map(context.delete)
-         return ManagedCache(context: context)
-     }
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-    @NSManaged var id: UUID
-    @NSManaged var imageDescription: String?
-    @NSManaged var location: String?
-    @NSManaged var url: URL
-    @NSManaged var cache: ManagedCache
-    
-    var local: LocalFeedImage {
-        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
-    }
-    
-    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        let managedFeed: [ManagedFeedImage] = localFeed.map {
-            let managed = ManagedFeedImage(context: context)
-            managed.id = $0.id
-            managed.imageDescription = $0.description
-            managed.location = $0.location
-            managed.url = $0.url
-            return managed
-        }
-        
-        return NSOrderedSet(array: managedFeed)
-    }
-}
